@@ -4,41 +4,24 @@ declare(strict_types=1);
 
 namespace Zipkin;
 
-use Zipkin\Propagation\B3;
-use Zipkin\Propagation\CurrentTraceContext;
-use Zipkin\Propagation\Propagation;
 use Zipkin\Reporter;
+use Zipkin\Propagation\Propagation;
+use Zipkin\Propagation\CurrentTraceContext;
+use Zipkin\Propagation\B3;
 
 final class DefaultTracing implements Tracing
 {
-    /**
-     * @var Tracer
-     */
-    private $tracer;
+    private Tracer $tracer;
 
-    /**
-     * @var Propagation
-     */
-    private $propagation;
+    private Propagation $propagation;
 
-    /**
-     * @var bool
-     */
-    private $isNoop;
+    private bool $isNoop;
 
-    /**
-     * @param Endpoint $localEndpoint
-     * @param Reporter $reporter
-     * @param Sampler $sampler
-     * @param bool $usesTraceId128bits
-     * @param CurrentTraceContext $currentTraceContext
-     * @param bool $isNoop
-     */
     public function __construct(
         Endpoint $localEndpoint,
         Reporter $reporter,
         Sampler $sampler,
-        $usesTraceId128bits,
+        bool $usesTraceId128bits,
         CurrentTraceContext $currentTraceContext,
         bool $isNoop,
         Propagation $propagation,
@@ -60,17 +43,11 @@ final class DefaultTracing implements Tracing
         $this->isNoop = $isNoop;
     }
 
-    /**
-     * @return Tracer
-     */
     public function getTracer(): Tracer
     {
         return $this->tracer;
     }
 
-    /**
-     * @return Propagation
-     */
     public function getPropagation(): Propagation
     {
         return $this->propagation;
@@ -80,7 +57,6 @@ final class DefaultTracing implements Tracing
      * When true, no recording is done and nothing is reported to zipkin. However, trace context is
      * still injected into outgoing requests.
      *
-     * @return bool
      * @see Span#isNoop()
      */
     public function isNoop(): bool
