@@ -18,10 +18,7 @@ namespace Zipkin\Propagation;
 
 final class CurrentTraceContext
 {
-    /**
-     * @var TraceContext|null
-     */
-    private $context;
+    private ?TraceContext $context;
 
     public function __construct(TraceContext $currentContext = null)
     {
@@ -30,8 +27,6 @@ final class CurrentTraceContext
 
     /**
      * Returns the current span context in scope or null if there isn't one.
-     *
-     * @return TraceContext|null
      */
     public function getContext(): ?TraceContext
     {
