@@ -4,23 +4,22 @@ declare(strict_types=1);
 
 namespace Zipkin\Reporters;
 
-use Psr\Log\LoggerInterface;
-use Zipkin\Recording\Span;
 use Zipkin\Reporter;
+use Zipkin\Recording\Span;
+use Psr\Log\LoggerInterface;
 
 final class Log implements Reporter
 {
-    /**
-     * @var LoggerInterface
-     */
-    private $logger;
+    private LoggerInterface $logger;
 
-    /**
-     * @param LoggerInterface $logger
-     */
-    public function __construct(LoggerInterface $logger)
-    {
+    private SpanSerializer $serializer;
+
+    public function __construct(
+        LoggerInterface $logger,
+        SpanSerializer $serializer = null
+    ) {
         $this->logger = $logger;
+        $this->serializer = $serializer ?? new JsonV2Serializer();
     }
 
     /**
@@ -28,8 +27,6 @@ final class Log implements Reporter
      */
     public function report(array $spans): void
     {
-        foreach ($spans as $span) {
-            $this->logger->info($span->__toString());
-        }
+        $this->logger->info($this->serializer->serialize($spans));
     }
 }
