@@ -2,13 +2,13 @@
 
 namespace ZipkinTests\Unit\Recording;
 
-use PHPUnit\Framework\TestCase;
+use Zipkin\Recording\Span;
+use Zipkin\Propagation\TraceContext;
+use Zipkin\Propagation\DefaultSamplingFlags;
 use Zipkin\Kind;
 use Zipkin\Endpoint;
-use Zipkin\Propagation\DefaultSamplingFlags;
-use Zipkin\Recording\Span;
+use PHPUnit\Framework\TestCase;
 use function Zipkin\Timestamp\now;
-use Zipkin\Propagation\TraceContext;
 
 final class SpanTest extends TestCase
 {
@@ -28,51 +28,5 @@ final class SpanTest extends TestCase
         $timestamp = now();
         $span->start($timestamp);
         $this->assertEquals($timestamp, $span->getTimestamp());
-    }
-
-    public function testConvertingSpanToArrayHasTheExpectedValues()
-    {
-        $spanId = 'e463f94de30144fa';
-        $traceId = 'e463f94de30144fa';
-        $parentId = 'e463f94de30144fb';
-
-        $context = TraceContext::create($traceId, $spanId, $parentId);
-
-        $localEndpoint = Endpoint::create('test_service_name', '127.0.0.1', null, 3333);
-        $span = Span::createFromContext($context, $localEndpoint);
-        $timestamp = now();
-
-        $span->start($timestamp);
-        $span->setKind(Kind\CLIENT);
-        $span->setName('test_name');
-        $span->annotate($timestamp, 'test_annotation');
-        $span->tag('test_key', 'test_value');
-        $span->finish($timestamp + 100);
-
-        $expectedSpanArray = [
-            'id' => $spanId,
-            'kind' => 'CLIENT',
-            'traceId' => $traceId,
-            'parentId' => $parentId,
-            'timestamp' => $timestamp,
-            'name' => 'test_name',
-            'duration' => 100,
-            'localEndpoint' => [
-                'serviceName' => 'test_service_name',
-                'ipv4' => '127.0.0.1',
-                'port' => 3333,
-            ],
-            'annotations' => [
-                [
-                    'value' => 'test_annotation',
-                    'timestamp' => $timestamp,
-                ],
-            ],
-            'tags' => [
-                'test_key' => 'test_value',
-            ],
-        ];
-
-        $this->assertEquals($expectedSpanArray, $span->toArray());
     }
 }
